Move Branch & Head Private's to own cpp
