+++ Libs/libMacGitverCore/RepoMan/Frontend/Head.cpp
@@ -21,90 +21,18 @@
 
 #include "RepoMan/Head.hpp"
 
-#include "RepoMan/Private/Dumper.hpp"
-
 #include "RepoMan/Data/HeadData.hpp"
 
 #include "RepoMan/Branch.hpp"
 #include "RepoMan/Repo.hpp"
 
 #include "libGitWrap/Repository.hpp"
-#include "libGitWrap/BranchRef.hpp"
 
 namespace RM
 {
 
-    namespace Internal
-    {
-
-        // -- HeadPrivate ----------------------------------------------------------------------- >8
-
-        HeadPrivate::HeadPrivate(Head* _pub, const Git::Repository& repo)
-            : BasePrivate(_pub)
-            , symbolicName()
-            , isDetached(repo.isHeadDetached())
-            , isUnborn(repo.isHeadUnborn())
-        {
-            Git::Result r;
-
-            if (isDetached) {
-                detachedId = repo.HEAD(r).objectId();
-            }
-            else {
-                symbolicName = repo.headBranchName(r);
-            }
-        }
-
-        ObjTypes HeadPrivate::objType() const
-        {
-            return ObjTypes::Head;
-        }
-
-        void HeadPrivate::dumpSelf(Internal::Dumper& dumper) const
-        {
-            dumper.addLine(QString(QStringLiteral("Head 0x%1 - %2"))
-                           .arg(quintptr(mPub),0,16)
-                           .arg(symbolicName));
-        }
-
-        bool HeadPrivate::refreshSelf()
-        {
-            Git::Repository repo = repository()->gitLoadedRepo();
-            Git::Result r;
-
-            isUnborn = repo.isHeadUnborn();
-            isDetached = repo.isHeadDetached();
-
-            if (isDetached) {
-                detachedId = repo.HEAD(r).objectId();
-            }
-            else {
-                symbolicName = repo.headBranchName(r);
-            }
-            return true;
-        }
-
-        QString HeadPrivate::displayName() const
-        {
-            return symbolicName;
-        }
-
-        QString HeadPrivate::objectTypeName() const
-        {
-            return QStringLiteral("Head");
-        }
-
-        bool HeadPrivate::inherits(ObjTypes type) const
-        {
-            return type == ObjTypes::Head || BasePrivate::inherits(type);
-        }
-
-    }
-
-    // -- Head ---------------------------------------------------------------------------------- >8
-
-    Head::Head(const Git::Repository& ref, Base* parent)
-        : Base(*new Internal::HeadPrivate(this, ref))
+    Head::Head(const Git::Repository& repo, Base* parent)
+        : Base(*new Internal::HeadPrivate(this, repo))
     {
         RM_D(Head);
         d->linkToParent(parent);
